design

+++ database/migrations/2020_03_18_085011_create_sizes_table.php
@@ -16,7 +16,7 @@ class CreateSizesTable extends Migration
         Schema::create('sizes', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->integer('weight')->default(1);
+            $table->float('weight')->default(1.0);
             $table->timestamps();
         });
     }
